requirements.txt
# CONCH Inference Dependencies
# Install with: pip install -r requirements.txt

# Core dependencies
h5py>=3.10.0
numpy>=1.24.0
pandas>=2.0.0
PyYAML>=6.0

# Deep learning
torch>=2.0.0
torchvision>=0.15.0

# For linear probe classification (optional)
scikit-learn>=1.3.0

# For image processing (if needed)
pillow>=10.0.0
tifffile>=2023.0.0

# For visualizations and GIFs
matplotlib>=3.7.0
seaborn>=0.12.0
imageio>=2.31.0
tqdm>=4.66.0

# For CONCH model (install from GitHub)
# git+https://github.com/mahmoodlab/CONCH.git
